refactor: reorganize folder structure
separate modules on their specific domains
