--- src/createLabeledUrl.ts
@@ -1,7 +1,0 @@
-export interface LabeledUrl {
-  label: string;
-  url: string;
-}
-
-export const createLabeledUrl = (label: string, url: string) =>
-  ({ label, url } as LabeledUrl);
